+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 
 
 ##
-## [0.3.1] 
+## [0.3.1] - 2019-05-23
 ### Changed
 - Bug fix: adding packages before all existing one works incorrect.